+++ packages/types/src/chain/parachain/Parachain.ts
@@ -10,6 +10,7 @@ export interface ParachainConstructorParams
   genesisHash: string;
   parachainId: number;
   ss58Format: number;
+  usesChainDecimals?: boolean;
   weight?: number;
   ws: string;
 }
@@ -23,6 +24,8 @@ export class Parachain extends Chain {
 
   readonly ss58Format: number;
 
+  readonly usesChainDecimals: boolean;
+
   readonly weight: number | undefined;
 
   readonly ws: string;
@@ -31,6 +34,7 @@ export class Parachain extends Chain {
     assetsData,
     genesisHash,
     parachainId,
+    usesChainDecimals,
     ss58Format,
     weight,
     ws,
@@ -46,6 +50,7 @@ export class Parachain extends Chain {
     this.genesisHash = genesisHash;
     this.parachainId = parachainId;
     this.ss58Format = ss58Format;
+    this.usesChainDecimals = !!usesChainDecimals;
     this.weight = weight;
     this.ws = ws;
   }