+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @moonbeam-network/xcm-sdk
 
+## 1.4.23
+
+### Patch Changes
+
+- [#179](https://github.com/moonbeam-foundation/xcm-sdk/pull/179) [`eb20aee`](https://github.com/moonbeam-foundation/xcm-sdk/commit/eb20aee9d9c9052074501421d0d442bd13022424) Thanks [@mmaurello](https://github.com/mmaurello)! - USDC wh integration with Moonbeam
+
+- Updated dependencies [[`eb20aee`](https://github.com/moonbeam-foundation/xcm-sdk/commit/eb20aee9d9c9052074501421d0d442bd13022424)]:
+  - @moonbeam-network/xcm-config@1.3.27
+  - @moonbeam-network/xcm-types@1.0.4
+  - @moonbeam-network/xcm-builder@1.0.13
+
+## 1.4.22
+
+### Patch Changes
+
+- Updated dependencies [[`6849ab3`](https://github.com/moonbeam-foundation/xcm-sdk/commit/6849ab3a5aa200d753849bc710d35fad4e2ac5fd)]:
+  - @moonbeam-network/xcm-config@1.3.26
+
 ## 1.4.21
 
 ### Patch Changes